A lot of small improvements.
- Fixed an issue that caused SNPs w p-values rounded to 0 to be excluded
from analysis.  If the p-value is 0, LDpred now attempts to use the SE
together with the effect size instead.

- Added an option that enables the user to always use SE together with
the effect size instead of p-value.

- The gibbs sampler now accounts for differing sample sizes per SNP, if
those are provided in the summary stats.  This seems to improve
robustness of the gibbs samler, and leads to better predictions in
simulations.  However, this option slows the Gibbs sampler a lot.

- It now estimates the LD score heritability per chromosome by default
(instead of partitioning it proportional to the number of SNPs.)

- Improved testing